Neural_Style_Transfer_Model/requirements.txt
#requirements
numpy==1.26.4
pandas==2.2.2
torch==2.2.2+cpu
torchvision==0.17.2+cpu
Pillow==10.3.0
streamlit==1.35.0
tqdm==4.66.4
-f https://download.pytorch.org/whl/cpu/torch_stable.html


